pyproject.toml
[project]
name = "recommendation-service"
version = "0.1.0"
description = ""
authors = [{ name = "andre", email = "andrepedro2004@hotmail.com" }]
requires-python = ">=3.11,<3.13"
readme = "README.md"
dependencies = [
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.12",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "redisvl>=0.3.9",
    "requests>=2.32.3",
    "strawberry-graphql>=0.265.1",
    "sentence-transformers"
]

[dependency-groups]
dev = [
    "pytest>=8.3.5,<9",
    "pytest-asyncio>=0.25.3,<0.26",
    "black>=25.1.0,<26",
    "isort>=6.0.1,<7",
    "watchfiles>=1.0.4,<2",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
